Fixing max value
This is a super hack. The reason I have to do this hack is because
the slider can't do a logarithmic scale. So although I want the max to
be lg(2^21 - 1), we don't have enough floating point precision to do this.

So the hack is the scale goes from 1 to 21^21, and then we just subtract 1
after getting the value from the slider.